requirements.txt
torch==2.5.1
typer==0.15.1
fastapi==0.115.6
uvicorn==0.34.0
timm==1.0.13
torchvision==0.20.1
torchaudio==2.5.1
lightning==2.5.0
gdown==5.2.0
scikit-image==0.25.0
matplotlib==3.10.0
omegaconf==2.3.0
hydra-core==1.3.2
opencv-python==4.10.0.84
wandb==0.19.4
black==24.10.0
isort==5.13.2
onnx==1.17.0
onnxruntime==1.20.1
onnxscript==0.1.0.dev20250117
lightning[extra]==2.5.0.post0
python-dotenv==1.0.1